[ENDIAN] Fix goldfish fb for bigendian targets
Swap bytes in the display buffer if the host endianess does not match
the target endianess. This fixes a problem with the wrong colour
being displayed in the text framebuffer console.
Not sure if the is the correct place to fix it.. It might be better to
have the kernel goldfish framebuffer driver swap the bytes.

Signed-off-by: Chris Dearman <chris@mips.com>